Refactor lore management tabs
